+++ app/Http/Controllers/VehicleController.php
@@ -5,17 +5,19 @@ namespace App\Http\Controllers;
 use App\Filters\VehicleFilters;
 use App\Models\Vehicle;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class VehicleController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('admin');
+        $this->middleware('auth');
     }
 
     public function index(Request $request, VehicleFilters $filters)
     {
+        Gate::authorize('edit-product');
         if ($request->wantsJson())
             return response()->json(Vehicle::filter($filters)->paginate(10));
         return view('dashboard.vehicle.index');
@@ -23,11 +25,13 @@ class VehicleController extends Controller
 
     public function create()
     {
+        Gate::authorize('edit-product');
         return view('dashboard.vehicle.create');
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('edit-product');
         $validated = $request->validate([
             'name' => 'required|string',
             'description' => 'string',
@@ -39,11 +43,13 @@ class VehicleController extends Controller
 
     public function edit(Vehicle $vehicle)
     {
+        Gate::authorize('edit-product');
         return view('dashboard.vehicle.edit', compact('vehicle'));
     }
 
     public function update(Vehicle $vehicle, Request $request)
     {
+        Gate::authorize('edit-product');
         $validated = $request->validate([
             'name' => 'required|string',
             'description' => 'string',
@@ -56,6 +62,7 @@ class VehicleController extends Controller
 
     public function destroy(Vehicle $vehicle)
     {
+        Gate::authorize('edit-product');
         Storage::delete($vehicle->splash);
         $vehicle->delete();
         return back();
